+++ DDM-OS-Reminder End-user Message.zsh
@@ -20,7 +20,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="1.4.0b4"
+scriptVersion="1.4.0"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"